+++ src/main/java/lotto/MyMoney.java
@@ -1,11 +1,11 @@
 package lotto;
 
 public class MyMoney {
-    private final int amount;
+    private final long amount;
 
     public MyMoney(String myMoney) {
         checkInputMoney(myMoney);
-        this.amount = Integer.parseInt(myMoney);
+        this.amount = Long.parseLong(myMoney);
     }
 
     private void checkIsZeroStart(String myMoney) {
@@ -14,21 +14,21 @@ public class MyMoney {
         }
     }
 
-    private int checkIsNumber(String myMoney) {
+    private long checkIsNumber(String myMoney) {
         try {
-            return Integer.parseInt(myMoney);
+            return Long.parseLong(myMoney);
         } catch (NumberFormatException exception) {
             throw new IllegalArgumentException("올바른 숫자를 입력해주세요. (문자가 포함된 입력)");
         }
     }
 
-    private void checkIsPlus(int moneyNum){
+    private void checkIsPlus(long moneyNum){
         if(moneyNum < 0){
             throw new IllegalArgumentException("올바를 숫자를 입력해주세요. (음수)");
         }
     }
 
-    private void checkIsDivide(int moneyNum) {
+    private void checkIsDivide(long moneyNum) {
         if (moneyNum % 1000 != 0) {
             throw new IllegalArgumentException("입력값은 1000으로 나누어 떨어져야 합니다.");
         }
@@ -36,12 +36,12 @@ public class MyMoney {
 
     private void checkInputMoney(String myMoney) {
         checkIsZeroStart(myMoney);
-        int moneyNum = checkIsNumber(myMoney);
+        long moneyNum = checkIsNumber(myMoney);
         checkIsPlus(moneyNum);
         checkIsDivide(moneyNum);
     }
 
-    public int getAmount() {
+    public long getAmount() {
         return this.amount;
     }
 }
